+++ src/UI/Components/ServiceInstanceForm/Components/suggestionResolvers.ts
@@ -14,6 +14,10 @@ export const resolveLabel = (
 /**
  * Resolves the value to submit for typed text: if it exactly matches a
  * suggestion's label, returns that suggestion's value; otherwise the text itself.
+ *
+ * Note: typing a string that exactly equals a label is deliberately treated as
+ * selecting that suggestion, so the underlying value is submitted rather than
+ * the literal text.
  */
 export const resolveValue = (
   suggestions: SuggestionValue[] | null | undefined,
